Move blog roll utilities to utilities folder
